Make AuthManagerAutoConfig configuration key more distinctive
The key is used in the AuthManagerFilterProviders hook so make it a
little more descriptive.

Note: deploying this will break in-progress login sessions, as
AuthManager does not like it when provider keys change mid-flight.
There will be some logspam for up to five minutes after deployment
("Authentication failed because of inconsistent provider array").
That's unavoidable, and low-impact (users will see a generic login
error; their next login attempt will start a new session and succeed).

Bug: T369180
Change-Id: I71427b2bb1890dc8c22844115dccbe8078a20568